refactor(review): remove helpers.py and move functions
