+++ homeassistant/components/flexmeasures/__init__.py
@@ -12,8 +12,11 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
 from .const import DOMAIN, FRBC_CONFIG
-from .helpers import get_from_option_or_config
-from .services import async_setup_services, async_unload_services
+from .services import (
+    async_setup_services,
+    async_unload_services,
+    get_from_option_or_config,
+)
 from .websockets import WebsocketAPIView
 
 _LOGGER = logging.getLogger(__name__)
