+++ homeassistant/components/flexmeasures/services.py
@@ -23,7 +23,6 @@ from .const import (
     SIGNAL_UPDATE_SCHEDULE,
 )
 from .exception import UndefinedCEMError, UnknownControlType
-from .helpers import get_from_option_or_config, time_ceil
 
 CHANGE_CONTROL_TYPE_SCHEMA = vol.Schema({vol.Optional("control_type"): str})
 
@@ -151,3 +150,23 @@ async def async_unload_services(hass: HomeAssistant) -> None:
     for service in SERVICES:
         if hass.services.has_service(DOMAIN, service["service"]):
             hass.services.async_remove(DOMAIN, service["service"])
+
+
+def time_mod(time, delta, epoch=None):
+    """From https://stackoverflow.com/a/57877961/13775459."""
+    if epoch is None:
+        epoch = datetime(1970, 1, 1, tzinfo=time.tzinfo)
+    return (time - epoch) % delta
+
+
+def time_ceil(time, delta, epoch=None):
+    """From https://stackoverflow.com/a/57877961/13775459."""
+    mod = time_mod(time, delta, epoch)
+    if mod:
+        return time + (delta - mod)
+    return time
+
+
+def get_from_option_or_config(key: str, entry: ConfigEntry):
+    """Get value from the options and, if not found, return the config value."""
+    return entry.options.get(key, entry.data.get(key))
